Rework timestamp prefix in console.
Previous implementation updated 'txt' with an array whose lifespan did
not exceed the scope of the if-block on lines 513-519. This is undefined
behavior.

The extra member 'prefix' in the console struct allows us to fill in a
timestamp and use this with every new line. This has the advantage that
it does not need to copy 'txt' and even handles multilines correctly
unlike the previous solution. Note however that it's unclear atm what
happens when con.linewidth < 9.

Also, move the early return of cl_noprint to the top.