Implement r_teleporterFlash cvar
r_teleporerFlash is a new cvar which allows more photosensitive users
to opt out of seeing a white flash every time they are in hyperspace
between a teleporter entrance and its exit. For now there is only a fade
to black option implemented alongside the fade to white original and
default option.

As a concept this is not new, DeFRaG has had a shader for this for a
long time and recently cnq3 (CPMA) added an engine cvar too. I have not
looked at other engines or mods for what they are doing. I chose the
same cvar name as cnq3 for cross-engine config compatibility and user
intuitivity as I do not see a reason to change it.

This cvar is future-proof and extendable as more styles are easy
to add with the cvar value being 2 or above. 0 for black, 1 for white,
2 and above for new funky variants.

These changes implement the cvar in renderer, renderer2 and renderervk.

Signed-off-by: drjaska <drjaska83@gmail.com>