server: don't delta from pre-map restart messages
During map restart, client->deltaActive is set to false for active clients via SV_ClientEnterWorld. This typically causes one non-delta snapshot to be sent, but as client moves come in, client->deltaActive is set back to true, and subsequent snapshots have delta enabled even though the source frame may be prior to the map restart.

This restores the behavior prior to 250179bac4f81cc10d70bf538b3185bb08d882ea where snapshots are non-delta until a post-restart frame is available, regardless of client commands. I don't expect this makes much difference in practice, but it is more consistent and perhaps a bit safer to stick to the original behavior in case of theoretical snapshot overflow issues.